src/backend/requirements.txt
# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9

# MCP
mcp[cli]==1.26.0

# Data Validation
pydantic==2.9.0
pydantic-settings==2.6.0

# HTTP Client
httpx==0.27.0

# Authentication
fastapi-login==1.0.2
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# Database
duckdb==0.10.0
sqlalchemy==2.0.35

# Cache
cachetools==5.3.3

# Encryption
cryptography==41.0.7

# LLM
anthropic==0.40.0
openai==1.54.0

# Utilities
python-dateutil==2.8.2
pyyaml==6.0.2

# Testing
pytest==8.3.0
pytest-asyncio==0.24.0
pytest-cov==5.0.0
